Fantom Subgraph Improvement
