--- Fantom/controller/subgraph.yaml
@@ -1,45 +1,0 @@
-specVersion: 0.0.2
-schema:
-  file: ./schema.graphql
-dataSources:
-  - kind: ethereum
-    name: Controller
-    network: fantom-testnet
-    source:
-      address: "0x3c141E6c5f6317Ce22d544c8B796ddeb6C6E6DEc"
-      abi: Controller
-      startBlock: 8524294
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.5
-      language: wasm/assemblyscript
-      entities:
-        - AdminChanged
-        - BeaconUpgraded
-        - CreateVest
-        - OwnershipTransferred
-        - ProjectInfo
-        - TransferWrapped
-        - Upgraded
-        - Withdraw
-      abis:
-        - name: Controller
-          file: ./abis/Controller.json
-      eventHandlers:
-        - event: AdminChanged(address,address)
-          handler: handleAdminChanged
-        - event: BeaconUpgraded(indexed address)
-          handler: handleBeaconUpgraded
-        - event: CreateVest(indexed address,address,address,uint256,uint256,address,string,bool)
-          handler: handleCreateVest
-        - event: OwnershipTransferred(indexed address,indexed address)
-          handler: handleOwnershipTransferred
-        - event: ProjectInfo(indexed address,string,address,uint256)
-          handler: handleProjectInfo
-        - event: TransferWrapped(address,indexed address,address,uint256)
-          handler: handleTransferWrapped
-        - event: Upgraded(indexed address)
-          handler: handleUpgraded
-        - event: Withdraw(indexed address,uint256,address)
-          handler: handleWithdraw
-      file: ./src/mapping.ts
