--- Fantom/master/subgraph.yaml
@@ -1,36 +1,0 @@
-specVersion: 0.0.2
-schema:
-  file: ./schema.graphql
-dataSources:
-  - kind: ethereum
-    name: Master
-    network: fantom-testnet
-    source:
-      address: "0xbfa83A2a2AcC83dFa90eFa2d0878054c6e37a5d6"
-      abi: Master
-      startBlock: 8524294
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.5
-      language: wasm/assemblyscript
-      entities:
-        - AdminChanged
-        - BeaconUpgraded
-        - OwnershipTransferred
-        - ProjectInfo
-        - Upgraded
-      abis:
-        - name: Master
-          file: ./abis/Master.json
-      eventHandlers:
-        - event: AdminChanged(address,address)
-          handler: handleAdminChanged
-        - event: BeaconUpgraded(indexed address)
-          handler: handleBeaconUpgraded
-        - event: OwnershipTransferred(indexed address,indexed address)
-          handler: handleOwnershipTransferred
-        - event: ProjectInfo(string,indexed address,string,string,address,uint256)
-          handler: handleProjectInfo
-        - event: Upgraded(indexed address)
-          handler: handleUpgraded
-      file: ./src/mapping.ts
