+++ Fantom/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "capx-liquid-fantom",
+  "license": "UNLICENSED",
+  "scripts": {
+    "codegen": "graph codegen",
+    "build": "graph build",
+    "deploy": "graph deploy --node https://api.thegraph.com/deploy/ shreyas3336/capx-liquid-fantom",
+    "create-local": "graph create --node http://localhost:8020/ shreyas3336/capx-liquid-fantom",
+    "remove-local": "graph remove --node http://localhost:8020/ shreyas3336/capx-liquid-fantom",
+    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 shreyas3336/capx-liquid-fantom"
+  },
+  "dependencies": {
+    "@graphprotocol/graph-cli": "0.29.0",
+    "@graphprotocol/graph-ts": "0.26.0"
+  }
+}
