--- Fantom/vesting/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "liquid-vesting-testnet-fantom",
-  "license": "UNLICENSED",
-  "scripts": {
-    "codegen": "graph codegen",
-    "build": "graph build",
-    "deploy": "graph deploy --node https://api.thegraph.com/deploy/ shreyas3336/liquid-vesting-testnet-fantom",
-    "create-local": "graph create --node http://localhost:8020/ shreyas3336/liquid-vesting-testnet-fantom",
-    "remove-local": "graph remove --node http://localhost:8020/ shreyas3336/liquid-vesting-testnet-fantom",
-    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 shreyas3336/liquid-vesting-testnet-fantom"
-  },
-  "dependencies": {
-    "@graphprotocol/graph-cli": "0.28.1",
-    "@graphprotocol/graph-ts": "0.26.0"
-  }
-}
