--- Matic (Polygon)/controller/schema.graphql
@@ -1,24 +1,0 @@
-type Project @entity {
-  id: ID!
-  projectOwnerAddress: Bytes! 
-  projectTokenAddress: Bytes!
-  projectTokenTicker: String!
-  projectTokenDecimal: BigInt!
-  derivatives: [Derivative!]! @derivedFrom(field: "projectID")
-}
-
-type Derivative @entity {
-  id: ID!
-  wrappedTokenTicker: String!
-  unlockTime: BigInt!
-  totalSupply: BigInt!
-  holders: [UserHoldings!]! @derivedFrom(field: "derivativeID")
-  projectID: Project!
-}
-
-type UserHoldings @entity {
-  id: ID!
-  address: Bytes!
-  tokenAmount: BigInt!
-  derivativeID: Derivative!
-}
