--- Matic (Polygon)/controller/subgraph.yaml
@@ -1,45 +1,0 @@
-specVersion: 0.0.2
-schema:
-  file: ./schema.graphql
-dataSources:
-  - kind: ethereum/contract
-    name: Controller
-    network: mumbai
-    source:
-      address: "0x73AfEFb5108AED20E80A564bAB866d6379051ECF"
-      abi: Controller
-      startBlock: 23019616
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.5
-      language: wasm/assemblyscript
-      entities:
-        - AdminChanged
-        - BeaconUpgraded
-        - CreateVest
-        - OwnershipTransferred
-        - ProjectInfo
-        - TransferWrapped
-        - Upgraded
-        - Withdraw
-      abis:
-        - name: Controller
-          file: ./abis/Controller.json
-      eventHandlers:
-        - event: AdminChanged(address,address)
-          handler: handleAdminChanged
-        - event: BeaconUpgraded(indexed address)
-          handler: handleBeaconUpgraded
-        - event: CreateVest(indexed address,address,address,uint256,uint256,address,string,bool)
-          handler: handleCreateVest
-        - event: OwnershipTransferred(indexed address,indexed address)
-          handler: handleOwnershipTransferred
-        - event: ProjectInfo(indexed address,string,address,uint256)
-          handler: handleProjectInfo
-        - event: TransferWrapped(address,indexed address,address,uint256)
-          handler: handleTransferWrapped
-        - event: Upgraded(indexed address)
-          handler: handleUpgraded
-        - event: Withdraw(indexed address,uint256,address)
-          handler: handleWithdraw
-      file: ./src/mapping.ts
