--- Matic (Polygon)/master/subgraph.yaml
@@ -1,36 +1,0 @@
-specVersion: 0.0.2
-schema:
-  file: ./schema.graphql
-dataSources:
-  - kind: ethereum/contract
-    name: Master
-    network: mumbai
-    source:
-      address: "0x9cCFc9B742856fE8E5dc520Da7D9377E1a10e77d"
-      abi: Master
-      startBlock: 23019611
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.5
-      language: wasm/assemblyscript
-      entities:
-        - AdminChanged
-        - BeaconUpgraded
-        - OwnershipTransferred
-        - ProjectInfo
-        - Upgraded
-      abis:
-        - name: Master
-          file: ./abis/Master.json
-      eventHandlers:
-        - event: AdminChanged(address,address)
-          handler: handleAdminChanged
-        - event: BeaconUpgraded(indexed address)
-          handler: handleBeaconUpgraded
-        - event: OwnershipTransferred(indexed address,indexed address)
-          handler: handleOwnershipTransferred
-        - event: ProjectInfo(string,indexed address,string,string,address,uint256)
-          handler: handleProjectInfo
-        - event: Upgraded(indexed address)
-          handler: handleUpgraded
-      file: ./src/mapping.ts
