--- Matic (Polygon)/vesting/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "liquid-vesting-mumbai-matic",
-  "license": "UNLICENSED",
-  "scripts": {
-    "codegen": "graph codegen",
-    "build": "graph build",
-    "deploy": "graph deploy --node https://api.thegraph.com/deploy/ shreyas3336/liquid-vesting-mumbai-matic",
-    "create-local": "graph create --node http://localhost:8020/ liquid-vesting-mumbai-matic",
-    "remove-local": "graph remove --node http://localhost:8020/ liquid-vesting-mumbai-matic",
-    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 liquid-vesting-mumbai-matic"
-  },
-  "dependencies": {
-    "@graphprotocol/graph-cli": "0.23.2",
-    "@graphprotocol/graph-ts": "0.23.1"
-  }
-}
