--- Matic (Polygon)/vesting/schema.graphql
@@ -1,22 +1,0 @@
-type Project @entity {
-  id: ID!
-  projectOwnerAddress: Bytes!
-  projectTokenAddress: Bytes!
-  derivatives: [Derivative!]! @derivedFrom(field: "projectID")
-}
-
-type Derivative @entity {
-  id: ID!
-  totalLockedSupply: BigInt!
-  holders: [UserHoldings!]! @derivedFrom(field: "derivativeID")
-  projectID: Project!
-}
-
-type UserHoldings @entity {
-  id: ID!
-  address: Bytes!
-  tokenAmount: BigInt!
-  unlockTime: BigInt!
-  vestID: BigInt!
-  derivativeID: Derivative!
-}
